Implementación de la interfaz de pasarelas de pago, se agrego Livewire y el formulario dinamico al comprar un plan, poder registrarte y pagar a través de Openpay. Se agrego el servicio de control UserService y el servicio de control de OpenpayService para controlar los cambios de usuarios local a usuarios en Openpay

+++ wave/database/migrations/2018_09_22_234251_create_announcement_user_table.php
@@ -14,8 +14,8 @@ class CreateAnnouncementUserTable extends Migration {
 	{
 		Schema::create('announcement_user', function(Blueprint $table)
 		{
-			$table->integer('announcement_id')->unsigned()->index();
-			$table->integer('user_id')->unsigned()->index();
+			$table->unsignedBigInteger('announcement_id')->unsigned()->index();
+			$table->unsignedBigInteger('user_id')->unsigned()->index();
 		});
 	}
